update changes for user authorization and authentication

+++ backend/controllers/goal.controller.js
@@ -1,9 +1,12 @@
 import asyncHandler from "express-async-handler";
 import { goalModel } from "../model/goal.model.js";
+import { userModel } from "../model/user.model.js";
 
 // get the all goals
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await goalModel.find();
+  const goals = await goalModel.find({
+    user: req.user.id,
+  });
 
   if (!goals) {
     res.status(400);
@@ -24,6 +27,7 @@ const createGoals = asyncHandler(async (req, res) => {
 
   const newGoal = {
     text: text,
+    user: req.user.id,
   };
 
   const result = await goalModel.create(newGoal);
@@ -35,9 +39,28 @@ const updateGoals = asyncHandler(async (req, res) => {
   const { id } = req.params;
   const { text } = req.body;
 
+  const goal = await goalModel.findById(id);
+
+  if (!goal) {
+    res.status(400);
+    throw new Error("Goal Not Found");
+  }
+
   if (!text) {
     res.status(400);
     throw new Error("Goal not Updated");
+  }
+
+  // check the user
+  if (!req.user) {
+    res.status(401);
+    throw new Error("User Not Found");
+  }
+
+  // /make sure only the logged user matches the goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User Not Authorized");
   }
 
   const updateGoal = { text: text };
@@ -51,9 +74,27 @@ const updateGoals = asyncHandler(async (req, res) => {
   res.status(200).json({ message: "Goal Updated Successfully" });
 });
 
-// delete the goal
 const deleteGoals = asyncHandler(async (req, res) => {
   const { id } = req.params;
+
+  const goal = await goalModel.findById(id);
+
+  if (!goal) {
+    res.status(400);
+    throw new Error("Goal Not Found");
+  }
+
+  // check the user
+  if (!req.user) {
+    res.status(401);
+    throw new Error("User Not Found");
+  }
+
+  // /make sure only the logged user matches the goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User Not Authorized");
+  }
 
   const result = await goalModel.findByIdAndDelete(id);
